requirements.txt
PyJWT==2.10.1
python-dotenv==1.0.1
Requests==2.32.3
rich==13.9.4
urllib3==2.3.0



#pip install -r requirements.txt